initial adjustment
