+++ internal/provider/deployment_resource.go
@@ -4,7 +4,6 @@ package provider
 
 import (
 	"context"
-	"encoding/json"
 	"fmt"
 	"net/http"
 
@@ -13,7 +12,6 @@ import (
 	"github.com/hashicorp/terraform-plugin-framework/path"
 	"github.com/hashicorp/terraform-plugin-framework/resource"
 	"github.com/hashicorp/terraform-plugin-framework/resource/schema"
-	"github.com/hashicorp/terraform-plugin-framework/resource/schema/booldefault"
 	"github.com/hashicorp/terraform-plugin-framework/resource/schema/defaults"
 	"github.com/hashicorp/terraform-plugin-framework/resource/schema/mapdefault"
 	"github.com/hashicorp/terraform-plugin-framework/resource/schema/planmodifier"
@@ -94,55 +92,7 @@ func (r *DeploymentResource) Schema(ctx context.Context, req resource.SchemaRequ
 				Description: "CEL expression to match job agents",
 			},
 		},
-		Blocks: map[string]schema.Block{
-			"argocd": schema.SingleNestedBlock{
-				Description: "ArgoCD job agent configuration",
-				Attributes: map[string]schema.Attribute{
-					"api_key":    schema.StringAttribute{Optional: true, Sensitive: true, Description: "ArgoCD API token"},
-					"server_url": schema.StringAttribute{Optional: true, Description: "ArgoCD server address"},
-					"template":   schema.StringAttribute{Optional: true, Description: "ArgoCD application template"},
-				},
-			},
-			"argo_workflow": schema.SingleNestedBlock{
-				Description: "Argo Workflow job agent configuration",
-				Attributes: map[string]schema.Attribute{
-					"api_key":        schema.StringAttribute{Optional: true, Sensitive: true, Description: "Argo Workflow API token"},
-					"server_url":     schema.StringAttribute{Optional: true, Description: "Argo Workflow server address"},
-					"template":       schema.StringAttribute{Optional: true, Description: "Argo Workflow application template"},
-					"name":           schema.StringAttribute{Optional: true, Description: "The name of the argo template to call"},
-					"webhook_secret": schema.StringAttribute{Optional: true, Sensitive: true, Description: "ArgoEvents webhook secret"},
-					"http_insecure":  schema.BoolAttribute{Optional: true, Computed: true, Description: "Allow insecure HTTP connections", Default: booldefault.StaticBool(false)},
-				},
-			},
-			"github": schema.SingleNestedBlock{
-				Description: "GitHub job agent configuration",
-				Attributes: map[string]schema.Attribute{
-					"installation_id": schema.Int64Attribute{Optional: true, Description: "GitHub app installation ID"},
-					"owner":           schema.StringAttribute{Optional: true, Description: "GitHub repository owner"},
-					"ref":             schema.StringAttribute{Optional: true, Description: "Git ref to run the workflow on"},
-					"repo":            schema.StringAttribute{Optional: true, Description: "GitHub repository name"},
-					"workflow_id":     schema.Int64Attribute{Optional: true, Description: "GitHub Actions workflow ID"},
-				},
-			},
-			"terraform_cloud": schema.SingleNestedBlock{
-				Description: "Terraform Cloud job agent configuration",
-				Attributes: map[string]schema.Attribute{
-					"address":               schema.StringAttribute{Optional: true, Description: "Terraform Cloud address"},
-					"organization":          schema.StringAttribute{Optional: true, Description: "Terraform Cloud organization name"},
-					"template":              schema.StringAttribute{Optional: true, Description: "Terraform Cloud workspace template"},
-					"token":                 schema.StringAttribute{Optional: true, Sensitive: true, Description: "Terraform Cloud API token"},
-					"trigger_run_on_change": schema.BoolAttribute{Optional: true, Description: "Whether to create a TFC run on dispatch"},
-				},
-			},
-			"test_runner": schema.SingleNestedBlock{
-				Description: "Test runner job agent configuration",
-				Attributes: map[string]schema.Attribute{
-					"delay_seconds": schema.Int64Attribute{Optional: true, Description: "Delay in seconds before resolving the job"},
-					"message":       schema.StringAttribute{Optional: true, Description: "Optional message to include in the job output"},
-					"status":        schema.StringAttribute{Optional: true, Description: "Final status to set"},
-				},
-			},
-		},
+		Blocks: jobAgentDispatchConfigBlocks(),
 	}
 }
 
@@ -153,23 +103,7 @@ func (r *DeploymentResource) ValidateConfig(ctx context.Context, req resource.Va
 		return
 	}
 
-	count := 0
-	if data.ArgoCD != nil {
-		count++
-	}
-	if data.ArgoWorkflow != nil {
-		count++
-	}
-	if data.GitHub != nil {
-		count++
-	}
-	if data.TerraformCloud != nil {
-		count++
-	}
-	if data.TestRunner != nil {
-		count++
-	}
-	if count > 1 {
+	if dispatchBlockCount(data.dispatchBlocks()) > 1 {
 		resp.Diagnostics.AddError(
 			"Invalid job agent configuration",
 			"Only one of argocd, argo_workflow, github, terraform_cloud, or test_runner can be set.",
@@ -388,119 +322,35 @@ type DeploymentResourceModel struct {
 	ResourceSelector types.String `tfsdk:"resource_selector"`
 	JobAgentSelector types.String `tfsdk:"job_agent_selector"`
 
-	ArgoCD         *DeploymentArgoCDModel       `tfsdk:"argocd"`
-	ArgoWorkflow   *DeploymentArgoWorkflowModel `tfsdk:"argo_workflow"`
-	GitHub         *DeploymentGitHubModel       `tfsdk:"github"`
-	TerraformCloud *DeploymentTFCModel          `tfsdk:"terraform_cloud"`
-	TestRunner     *DeploymentTestRunnerModel   `tfsdk:"test_runner"`
+	ArgoCD         *JobAgentDispatchArgoCDModel       `tfsdk:"argocd"`
+	ArgoWorkflow   *JobAgentDispatchArgoWorkflowModel `tfsdk:"argo_workflow"`
+	GitHub         *JobAgentDispatchGitHubModel       `tfsdk:"github"`
+	TerraformCloud *JobAgentDispatchTFCModel          `tfsdk:"terraform_cloud"`
+	TestRunner     *JobAgentDispatchTestRunnerModel   `tfsdk:"test_runner"`
 }
 
-type DeploymentArgoCDModel struct {
-	ApiKey    types.String `tfsdk:"api_key"`
-	ServerUrl types.String `tfsdk:"server_url"`
-	Template  types.String `tfsdk:"template"`
+func (m *DeploymentResourceModel) dispatchBlocks() JobAgentDispatchBlocks {
+	return JobAgentDispatchBlocks{
+		ArgoCD:         m.ArgoCD,
+		ArgoWorkflow:   m.ArgoWorkflow,
+		GitHub:         m.GitHub,
+		TerraformCloud: m.TerraformCloud,
+		TestRunner:     m.TestRunner,
+	}
 }
 
-type DeploymentArgoWorkflowModel struct {
-	ApiKey        types.String `tfsdk:"api_key"`
-	WebhookSecret types.String `tfsdk:"webhook_secret"`
-	ServerUrl     types.String `tfsdk:"server_url"`
-	Template      types.String `tfsdk:"template"`
-	Name          types.String `tfsdk:"name"`
-	HttpInsecure  types.Bool   `tfsdk:"http_insecure"`
-}
-
-type DeploymentGitHubModel struct {
-	InstallationId types.Int64  `tfsdk:"installation_id"`
-	Owner          types.String `tfsdk:"owner"`
-	Ref            types.String `tfsdk:"ref"`
-	Repo           types.String `tfsdk:"repo"`
-	WorkflowId     types.Int64  `tfsdk:"workflow_id"`
-}
-
-type DeploymentTFCModel struct {
-	Address            types.String `tfsdk:"address"`
-	Organization       types.String `tfsdk:"organization"`
-	Template           types.String `tfsdk:"template"`
-	Token              types.String `tfsdk:"token"`
-	TriggerRunOnChange types.Bool   `tfsdk:"trigger_run_on_change"`
-}
-
-type DeploymentTestRunnerModel struct {
-	DelaySeconds types.Int64  `tfsdk:"delay_seconds"`
-	Message      types.String `tfsdk:"message"`
-	Status       types.String `tfsdk:"status"`
+func (m *DeploymentResourceModel) setDispatchBlocks(b JobAgentDispatchBlocks) {
+	m.ArgoCD = b.ArgoCD
+	m.ArgoWorkflow = b.ArgoWorkflow
+	m.GitHub = b.GitHub
+	m.TerraformCloud = b.TerraformCloud
+	m.TestRunner = b.TestRunner
 }
 
 // deploymentJobAgentConfigFromModel extracts the typed block into a
 // map[string]interface{} suitable for the API's JobAgentConfig field.
 func deploymentJobAgentConfigFromModel(data *DeploymentResourceModel) *map[string]interface{} {
-	switch {
-	case data.ArgoCD != nil:
-		cfg := map[string]any{}
-		setStringIfSet(cfg, "apiKey", data.ArgoCD.ApiKey)
-		setStringIfSet(cfg, "serverUrl", data.ArgoCD.ServerUrl)
-		setStringIfSet(cfg, "template", data.ArgoCD.Template)
-		if len(cfg) == 0 {
-			return nil
-		}
-		return &cfg
-	case data.ArgoWorkflow != nil:
-		cfg := map[string]any{}
-		setStringIfSet(cfg, "apiKey", data.ArgoWorkflow.ApiKey)
-		setStringIfSet(cfg, "webhookSecret", data.ArgoWorkflow.WebhookSecret)
-		setStringIfSet(cfg, "serverUrl", data.ArgoWorkflow.ServerUrl)
-		setStringIfSet(cfg, "template", data.ArgoWorkflow.Template)
-		setStringIfSet(cfg, "name", data.ArgoWorkflow.Name)
-		if !data.ArgoWorkflow.HttpInsecure.IsNull() && !data.ArgoWorkflow.HttpInsecure.IsUnknown() {
-			cfg["httpInsecure"] = data.ArgoWorkflow.HttpInsecure.ValueBool()
-		}
-		if len(cfg) == 0 {
-			return nil
-		}
-		return &cfg
-	case data.GitHub != nil:
-		cfg := map[string]any{}
-		if !data.GitHub.InstallationId.IsNull() && !data.GitHub.InstallationId.IsUnknown() {
-			cfg["installationId"] = data.GitHub.InstallationId.ValueInt64()
-		}
-		setStringIfSet(cfg, "owner", data.GitHub.Owner)
-		setStringIfSet(cfg, "repo", data.GitHub.Repo)
-		setStringIfSet(cfg, "ref", data.GitHub.Ref)
-		if !data.GitHub.WorkflowId.IsNull() && !data.GitHub.WorkflowId.IsUnknown() {
-			cfg["workflowId"] = data.GitHub.WorkflowId.ValueInt64()
-		}
-		if len(cfg) == 0 {
-			return nil
-		}
-		return &cfg
-	case data.TerraformCloud != nil:
-		cfg := map[string]any{}
-		setStringIfSet(cfg, "address", data.TerraformCloud.Address)
-		setStringIfSet(cfg, "organization", data.TerraformCloud.Organization)
-		setStringIfSet(cfg, "template", data.TerraformCloud.Template)
-		setStringIfSet(cfg, "token", data.TerraformCloud.Token)
-		if !data.TerraformCloud.TriggerRunOnChange.IsNull() && !data.TerraformCloud.TriggerRunOnChange.IsUnknown() {
-			cfg["triggerRunOnChange"] = data.TerraformCloud.TriggerRunOnChange.ValueBool()
-		}
-		if len(cfg) == 0 {
-			return nil
-		}
-		return &cfg
-	case data.TestRunner != nil:
-		cfg := map[string]any{}
-		if !data.TestRunner.DelaySeconds.IsNull() && !data.TestRunner.DelaySeconds.IsUnknown() {
-			cfg["delaySeconds"] = data.TestRunner.DelaySeconds.ValueInt64()
-		}
-		setStringIfSet(cfg, "message", data.TestRunner.Message)
-		setStringIfSet(cfg, "status", data.TestRunner.Status)
-		if len(cfg) == 0 {
-			return nil
-		}
-		return &cfg
-	default:
-		return nil
-	}
+	return jobAgentDispatchConfigToMap(data.dispatchBlocks())
 }
 
 func setStringIfSet(cfg map[string]any, key string, val types.String) {
@@ -513,200 +363,10 @@ func setStringIfSet(cfg map[string]any, key string, val types.String) {
 // the API's JobAgentConfig map. It uses the prior state block type to decide
 // which block to populate so that reads are stable.
 func setDeploymentBlocksFromConfig(data *DeploymentResourceModel, config map[string]interface{}) {
-	blockType := deploymentBlockType(data)
-
-	// Preserve sensitive fields from prior state before clearing blocks.
-	priorArgoCD := data.ArgoCD
-	priorArgoWorkflow := data.ArgoWorkflow
-	priorTFC := data.TerraformCloud
-
-	data.ArgoCD = nil
-	data.ArgoWorkflow = nil
-	data.GitHub = nil
-	data.TerraformCloud = nil
-	data.TestRunner = nil
-
-	if len(config) == 0 {
-		return
-	}
-
-	if blockType == "" {
-		blockType = inferBlockTypeFromConfig(config)
-	}
-	if blockType == "" {
-		return
-	}
-
-	switch blockType {
-	case "argocd":
-		data.ArgoCD = &DeploymentArgoCDModel{
-			ApiKey:    stringValueOrNull(config["apiKey"]),
-			ServerUrl: stringValueOrNull(config["serverUrl"]),
-			Template:  stringValueOrNull(config["template"]),
-		}
-		if data.ArgoCD.ApiKey.IsNull() && priorArgoCD != nil && !priorArgoCD.ApiKey.IsNull() {
-			data.ArgoCD.ApiKey = priorArgoCD.ApiKey
-		}
-	case "argo_workflow":
-		data.ArgoWorkflow = &DeploymentArgoWorkflowModel{
-			ApiKey:        stringValueOrNull(config["apiKey"]),
-			WebhookSecret: stringValueOrNull(config["webhookSecret"]),
-			ServerUrl:     stringValueOrNull(config["serverUrl"]),
-			Template:      stringValueOrNull(config["template"]),
-			Name:          stringValueOrNull(config["name"]),
-			HttpInsecure:  boolValueOrNull(config["httpInsecure"]),
-		}
-		if data.ArgoWorkflow.ApiKey.IsNull() && priorArgoWorkflow != nil && !priorArgoWorkflow.ApiKey.IsNull() {
-			data.ArgoWorkflow.ApiKey = priorArgoWorkflow.ApiKey
-		}
-		if data.ArgoWorkflow.WebhookSecret.IsNull() && priorArgoWorkflow != nil && !priorArgoWorkflow.WebhookSecret.IsNull() {
-			data.ArgoWorkflow.WebhookSecret = priorArgoWorkflow.WebhookSecret
-		}
-	case "github":
-		gh := DeploymentGitHubModel{
-			InstallationId: types.Int64Null(),
-			Owner:          types.StringNull(),
-			Ref:            types.StringNull(),
-			Repo:           types.StringNull(),
-			WorkflowId:     types.Int64Null(),
-		}
-		if v, ok := config["installationId"]; ok && v != nil {
-			gh.InstallationId = types.Int64Value(toInt64(v))
-		}
-		if v, ok := config["owner"]; ok && v != nil && fmt.Sprint(v) != "" {
-			gh.Owner = types.StringValue(fmt.Sprint(v))
-		}
-		if v, ok := config["repo"]; ok && v != nil && fmt.Sprint(v) != "" {
-			gh.Repo = types.StringValue(fmt.Sprint(v))
-		}
-		if v, ok := config["ref"]; ok && v != nil && fmt.Sprint(v) != "" {
-			gh.Ref = types.StringValue(fmt.Sprint(v))
-		}
-		if v, ok := config["workflowId"]; ok && v != nil {
-			gh.WorkflowId = types.Int64Value(toInt64(v))
-		}
-		data.GitHub = &gh
-	case "terraform_cloud":
-		data.TerraformCloud = &DeploymentTFCModel{
-			Address:            stringValueOrNull(config["address"]),
-			Organization:       stringValueOrNull(config["organization"]),
-			Template:           stringValueOrNull(config["template"]),
-			Token:              stringValueOrNull(config["token"]),
-			TriggerRunOnChange: boolValueOrNull(config["triggerRunOnChange"]),
-		}
-		if data.TerraformCloud.Token.IsNull() && priorTFC != nil && !priorTFC.Token.IsNull() {
-			data.TerraformCloud.Token = priorTFC.Token
-		}
-	case "test_runner":
-		tr := DeploymentTestRunnerModel{
-			DelaySeconds: types.Int64Null(),
-			Message:      types.StringNull(),
-			Status:       types.StringNull(),
-		}
-		if v, ok := config["delaySeconds"]; ok && v != nil {
-			tr.DelaySeconds = types.Int64Value(toInt64(v))
-		}
-		if v, ok := config["message"]; ok && v != nil && fmt.Sprint(v) != "" {
-			tr.Message = types.StringValue(fmt.Sprint(v))
-		}
-		if v, ok := config["status"]; ok && v != nil && fmt.Sprint(v) != "" {
-			tr.Status = types.StringValue(fmt.Sprint(v))
-		}
-		data.TestRunner = &tr
-	}
-}
-
-type argoCDConfig struct {
-	ApiKey    string `json:"apiKey"`
-	ServerUrl string `json:"serverUrl"`
-	Template  string `json:"template"`
-}
-
-type argoWorkflowConfig struct {
-	ApiKey        string `json:"apiKey"`
-	WebhookSecret string `json:"webhookSecret"`
-	ServerUrl     string `json:"serverUrl"`
-	Template      string `json:"template"`
-	Name          string `json:"name"`
-	HttpInsecure  *bool  `json:"httpInsecure"`
-}
-
-type githubConfig struct {
-	InstallationId *int64 `json:"installationId"`
-	Owner          string `json:"owner"`
-	Ref            string `json:"ref"`
-	Repo           string `json:"repo"`
-	WorkflowId     *int64 `json:"workflowId"`
-}
-
-type terraformCloudConfig struct {
-	Address            string `json:"address"`
-	Organization       string `json:"organization"`
-	Template           string `json:"template"`
-	Token              string `json:"token"`
-	TriggerRunOnChange *bool  `json:"triggerRunOnChange"`
-}
-
-type testRunnerConfig struct {
-	DelaySeconds *int64 `json:"delaySeconds"`
-	Message      string `json:"message"`
-	Status       string `json:"status"`
-}
-
-func inferBlockTypeFromConfig(config map[string]interface{}) string {
-	data, err := json.Marshal(config)
-	if err != nil {
-		return ""
-	}
-
-	var gh githubConfig
-	_ = json.Unmarshal(data, &gh)
-	if gh.Owner != "" || gh.Repo != "" || gh.InstallationId != nil || gh.WorkflowId != nil {
-		return "github"
-	}
-
-	var tfc terraformCloudConfig
-	_ = json.Unmarshal(data, &tfc)
-	if tfc.Organization != "" || tfc.Address != "" || tfc.TriggerRunOnChange != nil {
-		return "terraform_cloud"
-	}
-
-	var tr testRunnerConfig
-	_ = json.Unmarshal(data, &tr)
-	if tr.DelaySeconds != nil || tr.Status != "" {
-		return "test_runner"
-	}
-
-	var aw argoWorkflowConfig
-	_ = json.Unmarshal(data, &aw)
-	if aw.WebhookSecret != "" || aw.HttpInsecure != nil || aw.Name != "" {
-		return "argo_workflow"
-	}
-
-	var ac argoCDConfig
-	_ = json.Unmarshal(data, &ac)
-	if ac.ServerUrl != "" || ac.Template != "" || ac.ApiKey != "" {
-		return "argocd"
-	}
-
-	return ""
-}
-
-func deploymentBlockType(data *DeploymentResourceModel) string {
-	switch {
-	case data.ArgoCD != nil:
-		return "argocd"
-	case data.ArgoWorkflow != nil:
-		return "argo_workflow"
-	case data.GitHub != nil:
-		return "github"
-	case data.TerraformCloud != nil:
-		return "terraform_cloud"
-	case data.TestRunner != nil:
-		return "test_runner"
-	default:
-		return ""
-	}
+	prior := data.dispatchBlocks()
+	var out JobAgentDispatchBlocks
+	setJobAgentDispatchBlocksFromConfig(&prior, &out, config)
+	data.setDispatchBlocks(out)
 }
 
 func stringValueOrNull(value interface{}) types.String {
